requirements.txt
Flask==0.10.1
Jinja2==2.7.1
MarkupSafe==0.18
Werkzeug==0.9.4
distribute==0.6.34
docker-py==0.2.0
greenlet==0.4.0
gunicorn==18.0
itsdangerous==0.23
meinheld==0.5.5
pymongo==2.6.2
pyres==1.4.2
redis==2.8.0
requests==1.2.3
setproctitle==1.1.7
six==1.4.1
